Move Link/Attribute test in the right package
