+++ leshan-core/src/test/java/org/eclipse/leshan/core/link/attributes/AttributeTest.java
@@ -1,4 +1,4 @@
-package org.eclipse.leshan.core.attributes;
+package org.eclipse.leshan.core.link.attributes;
 
 import static org.junit.Assert.*;
